requirements.txt
psycopg2-binary
sentence-transformers
pypdf
pandas
numpy
tqdm
llama-cpp-python
